refactor(client): extract student presence normalization
